Moved S3 related functions  to separate route
