+++ backend/server/index.js
@@ -1,10 +1,10 @@
 const dotenv = require('dotenv')
 dotenv.config()
 const express = require('express')
-const { formidable } = require('formidable')
-const { PassThrough } = require('stream')
+// const { formidable } = require('formidable')
+// const { PassThrough } = require('stream')
 
-const { listAllObjects, getUrl, getObject, uploadObject, putObject, deleteObject, } = require('./sdk/s3')
+// const { listAllObjects, getUrl, getObject, uploadObject, putObject, deleteObject, } = require('./sdk/s3')
 
 
 const app = express()
@@ -28,116 +28,104 @@ app.use('/', (req, res, next) => {
     next()
 })
 
+app.use('/s3', require('./routes/s3'))
+
 app.get('/', (req, res) => {
     res.send('Hello World!')
 })
 
 
-app.get('/list', (req, res) => {
-    listAllObjects().then(data => {
-        res.json(data)
-        // let data = JSON.stringify(r)
-        // res.writeHead(200, {
-        //     'Content-type': 'application/json'
-        // })
-        // res.write(data)
-        // res.end()
-    })
-})
+// app.get('/list', (req, res) => {
+//     listAllObjects().then(data => {
+//         res.json(data)
+//         // let data = JSON.stringify(r)
+//         // res.writeHead(200, {
+//         //     'Content-type': 'application/json'
+//         // })
+//         // res.write(data)
+//         // res.end()
+//     })
+// })
 
-app.get('/getUrl', async (req, res) => {
-    res.send('under developing')
-    // if (!req.query.objectKey) return res.status(400).json({ message: 'objectKey is required' })
-    // getUrl({
-    //     objectKey: req.query.objectKey
-    // }).then((url => {
-    //     res.send(url)
-    // })).catch(err => {
-    //     res.status(500).json({
-    //         message: err.message || 'Unexpected error occured'
-    //     })
-    // })
-})
+// app.get('/getObject', async (req, res) => {
+//     if (!req.query.objectKey) return res.status(400).json({ message: 'objectKey is required' })
+//     getObject({
+//         objectKey: req.query.objectKey
+//     }).then((data => {
+//         //data.Body is a readable stream
+//         const stream = data.Body
+//         stream.pipe(res)
+//     })).catch(err => {
+//         res.status(500).json({
+//             message: err.message || 'Unexpected error occured'
+//         })
+//     })
+// })
 
-app.get('/getObject', async (req, res) => {
-    if (!req.query.objectKey) return res.status(400).json({ message: 'objectKey is required' })
-    getObject({
-        objectKey: req.query.objectKey
-    }).then((data => {
-        //data.Body is a readable stream
-        const stream = data.Body
-        stream.pipe(res)
-    })).catch(err => {
-        res.status(500).json({
-            message: err.message || 'Unexpected error occured'
-        })
-    })
-})
+// app.post('/uploadObject', (req, res) => {
+//     // if (!req.query.objectKey) return res.status(400).json({ message: 'objectKey is required' })
 
-app.post('/uploadObject', (req, res) => {
-    // if (!req.query.objectKey) return res.status(400).json({ message: 'objectKey is required' })
+//     function fileWriteStreamHandler(file) {
+//         const pass = new PassThrough()
+//         uploadObject({
+//             objectKey: file.originalFilename,
+//             objectBody: pass
+//         }).then(data => {
+//             console.log(data)
+//             res.send(`${file.originalFilename} uploaded`)
+//         })
+//         return pass
+//     }
 
-    function fileWriteStreamHandler(file) {
-        const pass = new PassThrough()
-        uploadObject({
-            objectKey: file.originalFilename,
-            objectBody: pass
-        }).then(data => {
-            console.log(data)
-            res.send(`${file.originalFilename} uploaded`)
-        })
-        return pass
-    }
+//     const form = formidable({
+//         fileWriteStreamHandler: fileWriteStreamHandler
+//     })
+//     form.parse(req, async (err, fields, files) => {
+//         if (err) {
+//             res.status(500).json({
+//                 message: err.message || 'Unexpected error occured'
+//             })
+//             return
+//         }
 
-    const form = formidable({
-        fileWriteStreamHandler: fileWriteStreamHandler
-    })
-    form.parse(req, async (err, fields, files) => {
-        if (err) {
-            res.status(500).json({
-                message: err.message || 'Unexpected error occured'
-            })
-            return
-        }
+//     })
 
-    })
+// })
 
-})
+// app.post('/putObject', (req, res) => {
+//     // if (!req.query.objectKey) return res.status(400).json({ message: 'objectKey is required' })
+//     // const form = formidable({})
+//     // form.parse(req, (err, fields, files) => {
+//     //     if (err) {
+//     //         next(err)
+//     //         return
+//     //     }
+//     //     const objectKey = fields.objectKey
+//     //     const objectBody = files.objectBody[0]._writeStream.WriteStream
+//     //     console.log(objectBody)
+//     //     putObject({
+//     //         objectKey: objectKey,
+//     //         objectBody: objectBody
+//     //     }).then(data => {
+//     //         console.log(data)
+//     //     })
+//     // })
+// })
 
-app.post('/putObject', (req, res) => {
-    // if (!req.query.objectKey) return res.status(400).json({ message: 'objectKey is required' })
-    // const form = formidable({})
-    // form.parse(req, (err, fields, files) => {
-    //     if (err) {
-    //         next(err)
-    //         return
-    //     }
-    //     const objectKey = fields.objectKey
-    //     const objectBody = files.objectBody[0]._writeStream.WriteStream
-    //     console.log(objectBody)
-    //     putObject({
-    //         objectKey: objectKey,
-    //         objectBody: objectBody
-    //     }).then(data => {
-    //         console.log(data)
-    //     })
-    // })
-})
-
-app.post('/deleteObject', (req, res) => {
-    //somehow if the target object doesnt exist it still return as deleted
-    if (!req.query.objectKey) return res.status(400).json({ message: 'objectKey is required' })
-    deleteObject({
-        objectKey: req.query.objectKey
-    }).then(data => {
-        console.log(data)
-        res.send(`${req.query.objectKey} deleted`)
-    }).catch(err => {
-        res.status(500).json({
-            message: err.message || 'Unexpected error occured'
-        })
-    })
-})
+// app.post('/deleteObject', (req, res) => {
+//     //somehow if the target object doesnt exist it still return as deleted
+//     if (!req.query.objectKey) return res.status(400).json({ message: 'objectKey is required' })
+//     deleteObject({
+//         objectKey: req.query.objectKey
+//     }).then(data => {
+//         console.log(data)
+//         res.send(`${req.query.objectKey} deleted`)
+//     }).catch(err => {
+//         res.status(500).json({
+//             message: err.message || 'Unexpected error occured'
+//         })
+//     })
+// })
